// src/app/features/dashboard/pages/purchases/models/purchase-order-header.ts
export interface PurchaseOrderHeader {
  id: number;
  folio: string;
  purchaseOrderTypeId: number;
  purchaseOrderTypeName: string;
  businessUnitId: number;
  businessUnitName: string;
  costCenterId: number;
  costCenterName: string;
  requisitionHeader: number;
  supplierId: number;
  supplierName: string;
  supplierContactId: number;
  supplierContactName: string;
  paymentTypeId: number;
  paymentTypeName: string;
  paymentTermId: number;
  paymentTermName: string;
  statusId: number;
  statusName: string;
  currencyId: number;
  CurrencyCode: string;
  CurrencyName: string;
  LastCurrencyCode: string;
  LastTotal: number;
  plantId: number;
  plantName: string;
  LastPeriodStart: string;
  LastPeriodEnd: string;
  startPeriod: string;
  endPeriod: string;
  estimatedDate: string;
  previousAmount: string;
  subTotal: number;
  taxes: number;
  total: number;
  notes: string;
  observations: string;
  createBy: string;
  createdOn: string;
}